Add test cases for appProtocol
* pkg/router/router_test.go (harness): Add a client field.
(TestMain): Initialize the harness client field with the Kubernetes client set.
(TestConfigTemplate): Add test cases where the endpointslice associated with the
route specifies appProtocol with the recognized value "h2c" and unrecognized
values "unknown-value" and "kubernetes.io/h2c".  Verify that the recognized
value causes "proto h2" to be added to the backend server line and that the
unrecognized values do not.  Modify the test logic to apply an optional
mustCreateEndpointSlice from the test expectations.
(mustCreateRoute): Add targetServiceName field.
((mustCreateRoute).Apply): Initialize the route's spec.to.name field to the
targetServiceName value from the mustCreateRoute parameters if it is specified.
This new parameter isn't strictly required for the new test cases, but it makes
them more explicit and easier to read.
(mustCreateEndpointSlice): New type, used in the new test cases.
((mustCreateEndpointSlice).Apply): Create an endpointslice using the Kubernetes
API client set from the harness and the specified endpointslice name, service
name, and port appProtocol value from the mustCreateEndpointSlice parameters.
(mustCreateWithConfig): Add mustCreateEndpointSlice so that the new test cases
can use it.
(matchConfig): Add logic for parsing the "server" attribute, which the new test
cases use.