Add test cases for health check interval value
* pkg/router/router_test.go:
(TestConfigTemplate): Add test cases for route health check interval
annotation. Verify that the correct value is added to the backend server
line and that the values exceeding the maximum haproxy time value get clipped
to the max limit. Verify that invalid annotation values result in the default
health check interval value applied.
(MustCreateRoute): Add targetServiceWeight field
((MustCreateRoute).Apply): Initialize the route's spec.to.weight to the
targetServiceWeight value from the mustCreateRoute parameters if it is specified.
This parameter allows to set a non-zero weight for the service, making
the endpoint Active as a result.
(MustCreateEndpointSlice): Add addresses field
((MustCreateEndpointSlice).Apply): Initialize the endpoint's Addresses field
to the addresses array from the mustCreateEndpointSlice parameters if specified.
This enables having at least two endpoints for the route in the test case
and satisfy the conditions needed to configure the check inter in the backend
server line.