NE-1815: Add verifyhost to dynamic server slots for re-encrypt routes (DCM)
This commit introduces the "verifyhost" configuration setting to dynamic server slots during template rendering.
This ensures consistent behavior with static servers.

The "verifyhost" uses the FQDN of the primary service only,
as it's not feasible to predict which service endpoint will be used when alternate backends are available.
Consequently, dynamic servers for re-encrypt routes are not intended to handle alternate backend endpoints.
A reload is forced when endpoint changes occur for alternate backend services in a re-encrypt route.