OCPBUGS-55506: Prevent startup failures due to name resolution
When servers are listed using FQDN names rather than IP addresses,
HAProxy will attempt to resolve these names at startup. In the event
that a name cannot be resolve, HAProxy will fail. To overcome this,
use the 'init-addr' such that even if name resolution fails, HAProxy
will still start normally. Any server that fails name resolution will
be set in the DOWN state. Note that the key here is the 'none' keyword
used with 'init-addr'

Also add a new resolver to be used to periodically check that servers
specifed by are resolvable. Failure to resolve a server name will
result in the server being marked DOWN while successful name
resolution will result in the server being marked UP.

Also note that this option is being set via the 'default-server'
option in the 'defaults' section and therefore applied to all servers
listed in the configuration.