haproxy-config.template: Allow empty ciphers
Allow the empty value for TLS ciphers (TLSv1.2 and earlier) and cipher
suites (TLSv1.3), and use the empty value instead of setting a default
value.  Specifying the empty value means that the corresponding TLS
version cannot be used.

This commit fixes bug OCPBUGS-58040.

https://issues.redhat.com/browse/OCPBUGS-58040

* haproxy-config.template: Allow the empty value for ROUTER_CIPHERS and
ROUTER_CIPHERSUITES.  Rather than using a default value, pass an empty
value through to HAProxy (which in turn passes it through to OpenSSL).