OCPBUGS-49769: Validate the key and cert content before accepting
This change enforces the validation that a private key does not contain
any certificate/public key, and that a public key does not contain any
private key.

This is required because during the .pem file assembling for HAProxy,
in case a certificate is present on the private key field, it will be
ignored by the PEM sanitization function.

This lack of validation can cause a situation where the .pem file of
HAProxy will be created as private key -> CA chain -> leaf certificate,
which is invalid and rejected by HAProxy.

As the right order is private key -> leaf certificate -> CA Chain, and
the proper ordering of the public key/CA certificate is already guaranteed
by X509KeyPair(), we can reject the misusage of private keys or certificates
on the wrong fields, that can cause HAProxy to stop

Additionally, some tests were added to validate that PEM files exported
from a P12 and that contain bag attributes like friendlyNames are valid
and that the process of sanitization of these files will clean them correctly
and process they rest of the certificates and private keys