dynamicallyReplaceEndpoints: Check for idled svc
When dynamically replacing a route's endpoints, force a reload if the
new endpoints include an endpoint for an idled service.

Before this commit, the Dynamic Configuration Manager could replace an
active endpoint that had the health check enabled with the endpoint for
an idled service, leaving the health check enabled for the idled
service.  The health check could then cause the idled service to be
unidled.  This commit inhibits dynamic updates for a route that has an
idled endpoint; this means that the router follows the conventional
update logic, which does not enable health checks for idled services.

This commit fixes OCPBUGS-49908.

https://issues.redhat.com/browse/OCPBUGS-49908

* pkg/router/template/router.go (dynamicallyReplaceEndpoints): Log and
return false if the new endpoints include one for an idled service.