Do not install weak dependencies such as logrotate
When installing the haproxy-router image's package dependencies, specify
--setopt=install_weak_deps=0 to avoid install weak dependencies.

In particular, this commit avoids installing the rsyslog-logrotate
package, which is a weak dependency of the rsyslog package.  The
rsyslog-logrotate package requires the logrotate package; the router
does not need either of these packages.  Installing these unnecessary
packages enlarges the size and attack surface of the container image.

* hack/Dockerfile.debug:
* images/router/haproxy/Dockerfile:
* images/router/haproxy/Dockerfile.ocp:
* images/router/haproxy/Dockerfile.rhel: Specify the
--setopt=install_weak_deps=0 option for yum install.